# Optimize arithmetic for inexact real numbers