@jpellegrini 2024-07-24
@egallesio does this make sense?

I wish it was possible to not allocate reals when we do `a+b`, both being real, but that is currently not possible in STklos as I see...  But when they are different types, it is indeed possible!