@jpellegrini 2024-07-24
If this is OK, then maybe we can also do it for other types (complexes and rationals)?