@egallesio 2024-08-01
In particular, I have some regressions on  #586 and #587 that I don't understand. 
I'm afraid that this will need some more testing on some  rough corners .