@jpellegrini 2024-08-01
> Hi @jpellegrini,
> 
> I have done some work on this point since a couple of days and I have done it for all the types of numbers. The code is rather long and boring, and the function `convert` is completely gone. It is of course more efficient since we have less allocation, but a pretty side effect is that some special cases have disappeared (for instance, the multiplication of a real by a infinite), but there are still some problems to be resolved before I push my code.

That's great! :grinning:  I look forward to see this new code!