@egallesio 2024-08-31
@jpellegrini ,

Finally, I have committed my code... 
Now we should not  box useless numbers for the 4 operations: `+`, `-`, `*`, `/`. As I said before, the code is long and boring, but it should be OK. It also fixes some inconsistencies we had before (I had to fix some tests for that).

I close the PR, by hand